+++ src/containerapp/azext_containerapp/_help.py
@@ -186,11 +186,11 @@ helps['containerapp functions'] = """
 
 helps['containerapp functions list'] = """
     type: command
-    short-summary: List all functions in a container app or a specific revision.
+    short-summary: List all functions in a container app or a specific revision. (pass --revisionName parameter)
     long-summary: |
-        When revision name is not provided, lists functions from the active revision. If there are multiple active revisions, the API will return an error requiring you to specify which revision.
+        revisionName is required only if Container App active Revision Mode is setup in Multiple Revision Mode. (Default: Single Revision Mode)
     examples:
-    - name: List all functions in a container app (single active revision mode)
+    - name: List all functions in a container app. (single active revision mode)
       text: |
           az containerapp functions list -n my-containerapp -g MyResourceGroup
     - name: List all functions for a specific revision
@@ -200,11 +200,11 @@ helps['containerapp functions list'] = """
 
 helps['containerapp functions show'] = """
     type: command
-    short-summary: Show details of a specific function in a container app.
+    short-summary: Show details of a specific function in a container app or a specific revision within app. (pass --revisionName parameter)
     long-summary: |
-        When revision name is not provided, shows function from the active revision. If there are multiple active revisions, the API will return an error requiring you to specify which revision.
+        revisionName is required only if Container App active Revision Mode is setup in Multiple Revision Mode. (Default: Single Revision Mode)
     examples:
-    - name: Show details of a function in a container app (single active revision mode)
+    - name: Show details of a function in a container app. (single active revision mode)
       text: |
           az containerapp functions show -n my-containerapp -g MyResourceGroup --function-name MyFunction
     - name: Show details of a function for a specific revision